refactor(app): extract navigation/animation widgets out of shell_screen (Refs #563 phase: shell)
shell_screen.dart was 526 LOC — well over the 300 LOC ceiling tracked
by #563. Pull the two presentation surfaces (NavigationRail, bottom
bar) and the canonical destination-resolution helper into a sibling
`lib/app/shell/` package so the parent shell only owns animation +
navigation orchestration.

- shell/shell_nav_item.dart  — `ShellNavItem` data class + `ShellBounceIcon`
  (the icon-bounce animation is shared by both surfaces)
- shell/shell_nav_rail.dart  — `ShellNavRail` (was `_AdaptiveNavigationRail`)
- shell/shell_bottom_bar.dart — `ShellBottomBar` (was `_AnimatedNavBar`)
- shell/shell_destinations.dart — `resolveShellDestinations()` pure
  function + `kConsumptionBranchIndex` constant; covers the #893
  "drop Conso slot when no vehicle" rule

Behavior is unchanged. The #893 (Conso vehicle gating) and #595 (Hero
preservation via go_router StatefulShellRoute) semantics are
preserved verbatim — only the file boundaries moved.

Result: every file ≤ 300 LOC.
  shell_screen.dart       281
  shell_bottom_bar.dart   118
  shell_destinations.dart  77
  shell_nav_item.dart      71
  shell_nav_rail.dart      79

Tests: added `test/app/shell/shell_destinations_test.dart` covering
the pure resolver (5-tab default, 4-tab gated, branch-for-slot
mapping, Settings-rightmost invariant). All 4 existing shell tests
(shell_screen_test, shell_screen_responsive_test,
shell_consumption_tab_test, shell_nav_vehicle_gating_test) still
pass against the extracted code.
